fix(core-tests): 消除 GameContextTests 警告
- 拆分 TestArchitecture 与 TestArchitectureContext 到独立测试文件以消除 MA0048
- 修复 TestArchitectureContext 的可空签名以匹配 IArchitectureContext 契约
- 补充测试桩公开成员的 XML 文档并保持 GameContextTests 行为不变
